---------------------------------------------------------------------------------------------------------------------- Updates to Text Summarization from Dense Subgraph of Recursive Gloss Overlap graph of a Text ---------------------------------------------------------------------------------------------------------------------- (*) import matplotlib commented in RGO classifier (*) New clause added to Text Summarization: This finds the common path between 2 word vertices synsets by inheriting an existing WordNet code for shortest wordnet path distance and constructs sentences by a sliding window for each successive pair of intermediate vertices in this common path. This creates a deeper profund summary than least_common_hypernyms().Sentences thus created are sorted based on a relevance score to the actual text - size of intersection similar to Lesk WSD. Only top relevant sentences can be chosen as required. (*) Logs for this have been committed to testlogs/.
It has to be noted that, dense subgraph traversal and writing unguided summary sans training data mimicks human recursive
comprehension and looks non-conventional for human reading. Presently only hypernym (IS A) relation is used for connectives.
More comprehensive humane-looking summary can be created by Meronyms (HAS A) and Holonyms(IS PART OF) in WordNet API. ConceptNet 5 which is a better semantic framework for finding relations between concepts could do well in this context. But
presently there is no python API for ConceptNet5 and data has to be queried as RESTful JSON objects.